Renamed cf target package
